Split up main unit quantity and symbol
