+++ src/api/swapTypes.ts
@@ -1,23 +1,7 @@
-import { toBitcoin } from "satoshi-bitcoin-ts";
-import { fromWei } from "web3-utils";
-
 export interface Asset {
   name: string;
   quantity: string;
   token_contract?: string;
-}
-
-export function toMainUnit(asset: Asset) {
-  switch (asset.name) {
-    case "ether":
-      return fromWei(asset.quantity, "ether") + " ETH";
-    case "bitcoin":
-      return toBitcoin(asset.quantity, true) + " BTC";
-    case "erc20":
-      return asset.quantity + " ERC20";
-    default:
-      return asset.quantity + " " + asset.name;
-  }
 }
 
 export interface Ledger {
